+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # What's New with Infor Design System
 
+## v4.8.1
+
+### v4.8.1 Fixes
+
+- `[Color]` Added an importable ui.config.color-palette.js to use in the components. ([#478](https://github.com/infor-design/enterprise/issues/478))
+- `[Mixins]` Added flex mixins like tailwind. ([#478](https://github.com/infor-design/enterprise/issues/478))
+
 ## v4.8.0
 
 ### v4.8.0 Fixes
@@ -7,7 +14,7 @@
 - `[Colors]` Changed the color names for alerts and updated them for themes. ([#478](https://github.com/infor-design/enterprise/issues/478))
 - `[Mixins]` Added border radius mixins for every side. ([#478](https://github.com/infor-design/enterprise/issues/478))
 - `[Mixins]` Changed `ml mr mx pl pr px` to use inline-start/end for better RTL. ([#478](https://github.com/infor-design/enterprise/issues/478))
-- `[Tokens]` Added an importable fontSize.js to use in the components. ([#478](https://github.com/infor-design/enterprise/issues/478))
+- `[Tokens]` Added an importable ui.config.font-size.js to use in the components. ([#478](https://github.com/infor-design/enterprise/issues/478))
 
 ## v4.7.10